feat/refactor: refactored controller to work with new promise based connection and implemented authorization

+++ backend/src/controllers/forum-controller.ts
@@ -1,69 +1,90 @@
 import type { Request, Response } from "express"
 import { connection } from "../utils/db-connection"
-import { handleQueryOutput } from "../utils/query-handling"
+import { handleAuthorization, handleUser } from "../utils/auth"
 
 export const addComment = async (req: Request, res: Response) => {
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+
     await connection.execute(
         `INSERT INTO comments (gameId, userId, commentBody, commentTimeStamp) VALUES (?, ?, ?, NOW())`,
-        [req.body["gameId"], req.body["userId"], req.body["commentBody"]],
-        handleQueryOutput(201, res)
+        [req.body["gameId"], req.body["userId"], req.body["commentBody"]]
     )
+    res.status(201).send("Comment added successfully");
 };
 
 export const listForums = async (req: Request, res: Response) => {
-    await connection.execute(
+    const [forums] = await connection.execute(
         `SELECT gameId, gameName, gameSmallBannerName, gameSmallBannerBin FROM games`,
-        [],
-        handleQueryOutput(200, res)
+        []
     )
+    res.status(200).json(forums)
 };
 
 export const listForumsAsYouType = async (req: Request, res: Response) => {
-    await connection.execute(
+    const [forums] = await connection.execute(
         `SELECT gameId, gameName, gameSmallBannerBin, gameSmallBannerName
         FROM games
         WHERE gameName LIKE CONCAT('%', ?, '%') or 
         gameAlternateName like CONCAT('%', ?, '%')
         ORDER BY gameName DESC
         LIMIT 15`,
-        [req.params["partialName"], req.params["partialName"]],
-        handleQueryOutput(200, res)
+        [req.params["partialName"], req.params["partialName"]]
     )
+    res.status(200).json(forums)
 };
 
 export const listForumsMatching = async (req: Request, res: Response) => {
-    await connection.execute(
+    const [forums] = await connection.execute(
         `SELECT gameId, gameName, gameSmallBannerBin, gameSmallBannerName
         FROM games
         WHERE gameName LIKE CONCAT('%', ?, '%') or 
         gameAlternateName like CONCAT('%', ?, '%')
         ORDER BY gameName DESC`,
-        [req.params["partialName"], req.params["partialName"]],
-        handleQueryOutput(200, res)
+        [req.params["partialName"], req.params["partialName"]]
     )
+    res.status(200).json(forums)
 };
 
 export const listCommentsOfGame = async (req: Request, res: Response) => {
-    await connection.execute(
+    const [comments] = await connection.execute(
         `SELECT commentBody, commentTimeStamp, commentWasEdited, userUsername, userIconBin, userIconName 
         FROM comments as c join users as u ON c.userId = u.userId WHERE gameId = ? ORDER BY commentTimeStamp DESC`,
-        [req.params['gameId']],
-        handleQueryOutput(200, res)
+        [req.params['gameId']]
     )
+    res.status(200).json(comments)
 };
 
 export const updateComment = async (req: Request, res: Response) => {
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+    if(!handleAuthorization(res, user, req.body["userId"], false)) {
+        return;
+    }
+
     await connection.execute(
         `INSERT INTO comments (commentBody, commentWasEdited) VALUES (?, true) WHERE commentId = ?`,
-        [req.body["commentBody"], req.params["commentId"]],
-        handleQueryOutput(200, res)
+        [req.body["commentBody"], req.params["commentId"]]
     )
+    res.status(200).send("Comment updated successfully");
 };
 
 export const deleteComment = async (req: Request, res: Response) => {
+    const user = await handleUser(req, res, ["user", "admin", "master"]);
+    if (!user) {
+        return;
+    }
+    if(!handleAuthorization(res, user, req.body["userId"], true)) {
+        return;
+    }
+
     await connection.execute(
         `DELETE FROM comments WHERE commentId = ?`,
-        [req.params['commentId']],
-        handleQueryOutput(200, res)
+        [req.params['commentId']]
     )
+    res.status(200).send("Comment deleted successfully");
 };
